segfaulting, but log keys integrated into main
